+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 - Global `--no-sync` flag to not run any sync on command
 - `awless show policy-name/policy-id` now displays the current policy Document (in JSON).
 
+### AWS Services
+
+- Update IAM policies, to add statements with `awless update policy`
+
 ### Fixes
 - `awless ssh --through`: no reusing same conn to avoid EOF. Bug: only first user (amazonlinux) was successful (usually ec2-user) !!
 - `awless ssh --through`: on new proxy client catching error that where shadowed